Swap naming of error and system log on Dev Info panel
