+++ src/data/error_log.ts
@@ -1,4 +1,13 @@
 import { HomeAssistant } from "../types";
 
+export interface LoggedError {
+  message: string;
+  level: string;
+  source: string;
+  // unix timestamp in seconds
+  timestamp: number;
+  exception: string;
+}
+
 export const fetchErrorLog = (hass: HomeAssistant) =>
-  hass.callApi<string>("GET", "error_log");
+  hass.callApi<LoggedError[]>("GET", "error/all");
