+++ src/data/system_log.ts
@@ -1,13 +1,4 @@
 import { HomeAssistant } from "../types";
 
-export interface LoggedError {
-  message: string;
-  level: string;
-  source: string;
-  // unix timestamp in seconds
-  timestamp: number;
-  exception: string;
-}
-
 export const fetchSystemLog = (hass: HomeAssistant) =>
-  hass.callApi<LoggedError[]>("GET", "error/all");
+  hass.callApi<string>("GET", "error_log");
