+++ src/panels/dev-info/error-log-card.ts
@@ -6,67 +6,142 @@ import {
   PropertyDeclarations,
   TemplateResult,
 } from "lit-element";
+import "@polymer/paper-card/paper-card";
 import "@polymer/paper-icon-button/paper-icon-button";
-import "@polymer/paper-button/paper-button";
-
+import "@polymer/paper-item/paper-item-body";
+import "@polymer/paper-item/paper-item";
+import "@polymer/paper-spinner/paper-spinner";
+import "../../components/buttons/ha-call-service-button";
+import "../../components/buttons/ha-progress-button";
 import { HomeAssistant } from "../../types";
-import { fetchErrorLog } from "../../data/error_log";
+import { LoggedError, fetchErrorLog } from "../../data/error_log";
+import formatDateTime from "../../common/datetime/format_date_time";
+import formatTime from "../../common/datetime/format_time";
+import { showErrorLogDetailDialog } from "./show-dialog-error-log-detail";
+
+const formatLogTime = (date, language: string) => {
+  const today = new Date().setHours(0, 0, 0, 0);
+  const dateTime = new Date(date * 1000);
+  const dateTimeDay = new Date(date * 1000).setHours(0, 0, 0, 0);
+
+  return dateTimeDay < today
+    ? formatDateTime(dateTime, language)
+    : formatTime(dateTime, language);
+};
 
 class ErrorLogCard extends LitElement {
   public hass?: HomeAssistant;
-  private _errorLog?: string;
+  private _items?: LoggedError[];
 
   static get properties(): PropertyDeclarations {
     return {
       hass: {},
-      _errorLog: {},
+      _items: {},
     };
   }
 
   protected render(): TemplateResult | void {
     return html`
-      <p class="error-log-intro">
-        ${this._errorLog
-          ? html`
-              <paper-icon-button
-                icon="hass:refresh"
-                @click=${this._refreshErrorLog}
-              ></paper-icon-button>
-            `
-          : html`
-              <paper-button raised @click=${this._refreshErrorLog}>
-                Load Full Home Assistant Log
-              </paper-button>
-            `}
-      </p>
-      <div class="error-log">${this._errorLog}</div>
+      <div class="error-log">
+        <paper-card heading="Error Log">
+          ${this._items === undefined
+            ? html`
+                <div class="loading-container">
+                  <paper-spinner active></paper-spinner>
+                </div>
+              `
+            : html`
+                ${this._items.length === 0
+                  ? html`
+                      <div class="card-content">There are no new issues!</div>
+                    `
+                  : this._items.map(
+                      (item) => html`
+                        <paper-item @click=${this._openLog} .logItem=${item}>
+                          <paper-item-body two-line>
+                            <div class="row">
+                              ${item.message}
+                            </div>
+                            <div secondary>
+                              ${formatLogTime(
+                                item.timestamp,
+                                this.hass!.language
+                              )}
+                              ${item.source} (${item.level})
+                            </div>
+                          </paper-item-body>
+                        </paper-item>
+                      `
+                    )}
+
+                <div class="card-actions">
+                  <ha-call-service-button
+                    .hass=${this.hass}
+                    domain="system_log"
+                    service="clear"
+                    >Clear</ha-call-service-button
+                  >
+                  <ha-progress-button @click=${this._fetchData}
+                    >Refresh</ha-progress-button
+                  >
+                </div>
+              `}
+        </paper-card>
+      </div>
     `;
+  }
+
+  protected firstUpdated(changedProps): void {
+    super.firstUpdated(changedProps);
+    this._fetchData();
+    this.addEventListener("hass-service-called", (ev) =>
+      this.serviceCalled(ev)
+    );
+  }
+
+  protected serviceCalled(ev): void {
+    // Check if this is for us
+    if (ev.detail.success && ev.detail.domain === "system_log") {
+      // Do the right thing depending on service
+      if (ev.detail.service === "clear") {
+        this._items = [];
+      }
+    }
+  }
+
+  private async _fetchData(): Promise<void> {
+    this._items = undefined;
+    this._items = await fetchErrorLog(this.hass!);
+  }
+
+  private _openLog(ev: Event): void {
+    const item = (ev.currentTarget as any).logItem;
+    showErrorLogDetailDialog(this, { item });
   }
 
   static get styles(): CSSResult {
     return css`
-      .error-log-intro {
-        text-align: center;
-        margin: 16px;
+      paper-card {
+        display: block;
+        padding-top: 16px;
       }
 
-      paper-icon-button {
-        float: right;
+      paper-item {
+        cursor: pointer;
       }
 
       .error-log {
-        @apply --paper-font-code)
-          clear: both;
-        white-space: pre-wrap;
         margin: 16px;
+        border-top: 1px solid var(--light-primary-color);
+        padding-top: 16px;
+      }
+
+      .loading-container {
+        @apply --layout-vertical;
+        @apply --layout-center-center;
+        height: 100px;
       }
     `;
-  }
-
-  private async _refreshErrorLog(): Promise<void> {
-    this._errorLog = "Loading error log…";
-    const log = await fetchErrorLog(this.hass!);
-    this._errorLog = log || "No errors have been reported.";
   }
 }
