+++ src/panels/dev-info/system-log-card.ts
@@ -6,142 +6,67 @@ import {
   PropertyDeclarations,
   TemplateResult,
 } from "lit-element";
-import "@polymer/paper-card/paper-card";
 import "@polymer/paper-icon-button/paper-icon-button";
-import "@polymer/paper-item/paper-item-body";
-import "@polymer/paper-item/paper-item";
-import "@polymer/paper-spinner/paper-spinner";
-import "../../components/buttons/ha-call-service-button";
-import "../../components/buttons/ha-progress-button";
+import "@polymer/paper-button/paper-button";
+
 import { HomeAssistant } from "../../types";
-import { LoggedError, fetchSystemLog } from "../../data/system_log";
-import formatDateTime from "../../common/datetime/format_date_time";
-import formatTime from "../../common/datetime/format_time";
-import { showSystemLogDetailDialog } from "./show-dialog-system-log-detail";
-
-const formatLogTime = (date, language: string) => {
-  const today = new Date().setHours(0, 0, 0, 0);
-  const dateTime = new Date(date * 1000);
-  const dateTimeDay = new Date(date * 1000).setHours(0, 0, 0, 0);
-
-  return dateTimeDay < today
-    ? formatDateTime(dateTime, language)
-    : formatTime(dateTime, language);
-};
+import { fetchSystemLog } from "../../data/system_log";
 
 class SystemLogCard extends LitElement {
   public hass?: HomeAssistant;
-  private _items?: LoggedError[];
+  private _systemLog?: string;
 
   static get properties(): PropertyDeclarations {
     return {
       hass: {},
-      _items: {},
+      _systemLog: {},
     };
   }
 
   protected render(): TemplateResult | void {
     return html`
-      <div class="system-log-intro">
-        <paper-card>
-          ${this._items === undefined
-            ? html`
-                <div class="loading-container">
-                  <paper-spinner active></paper-spinner>
-                </div>
-              `
-            : html`
-                ${this._items.length === 0
-                  ? html`
-                      <div class="card-content">There are no new issues!</div>
-                    `
-                  : this._items.map(
-                      (item) => html`
-                        <paper-item @click=${this._openLog} .logItem=${item}>
-                          <paper-item-body two-line>
-                            <div class="row">
-                              ${item.message}
-                            </div>
-                            <div secondary>
-                              ${formatLogTime(
-                                item.timestamp,
-                                this.hass!.language
-                              )}
-                              ${item.source} (${item.level})
-                            </div>
-                          </paper-item-body>
-                        </paper-item>
-                      `
-                    )}
-
-                <div class="card-actions">
-                  <ha-call-service-button
-                    .hass=${this.hass}
-                    domain="system_log"
-                    service="clear"
-                    >Clear</ha-call-service-button
-                  >
-                  <ha-progress-button @click=${this._fetchData}
-                    >Refresh</ha-progress-button
-                  >
-                </div>
-              `}
-        </paper-card>
-      </div>
+      <p class="system-log-intro">
+        ${this._systemLog
+          ? html`
+              <paper-icon-button
+                icon="hass:refresh"
+                @click=${this._refreshSystemLog}
+              ></paper-icon-button>
+            `
+          : html`
+              <paper-button raised @click=${this._refreshSystemLog}>
+                Load Full Home Assistant Log
+              </paper-button>
+            `}
+      </p>
+      <div class="system-log">${this._systemLog}</div>
     `;
-  }
-
-  protected firstUpdated(changedProps): void {
-    super.firstUpdated(changedProps);
-    this._fetchData();
-    this.addEventListener("hass-service-called", (ev) =>
-      this.serviceCalled(ev)
-    );
-  }
-
-  protected serviceCalled(ev): void {
-    // Check if this is for us
-    if (ev.detail.success && ev.detail.domain === "system_log") {
-      // Do the right thing depending on service
-      if (ev.detail.service === "clear") {
-        this._items = [];
-      }
-    }
-  }
-
-  private async _fetchData(): Promise<void> {
-    this._items = undefined;
-    this._items = await fetchSystemLog(this.hass!);
-  }
-
-  private _openLog(ev: Event): void {
-    const item = (ev.currentTarget as any).logItem;
-    showSystemLogDetailDialog(this, { item });
   }
 
   static get styles(): CSSResult {
     return css`
-      paper-card {
-        display: block;
-        padding-top: 16px;
-      }
-
-      paper-item {
-        cursor: pointer;
-      }
-
       .system-log-intro {
+        text-align: center;
         margin: 16px;
-        border-top: 1px solid var(--light-primary-color);
-        padding-top: 16px;
       }
 
-      .loading-container {
-        @apply --layout-vertical;
-        @apply --layout-center-center;
-        height: 100px;
+      paper-icon-button {
+        float: right;
+      }
+
+      .system-log {
+        @apply --paper-font-code)
+          clear: both;
+        white-space: pre-wrap;
+        margin: 16px;
       }
     `;
+  }
+
+  private async _refreshSystemLog(): Promise<void> {
+    this._systemLog = "Loading full system log…";
+    const log = await fetchSystemLog(this.hass!);
+    this._systemLog = log || "No log at this time.";
   }
 }
